requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
pinecone-client==2.2.4
weaviate-client==4.4.3
faiss-cpu==1.7.4
chromadb==0.4.18
elasticsearch==8.10.1
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
requests==2.31.0
pydantic==2.5.2 